refactor(mobile): move ContactItem to separate file
