+++ apps/mobile/src/features/AddressBook/List/components/List/index.ts
@@ -1,3 +1,4 @@
 import { ContactListItems } from './ContactListItems'
 export { ContactListItems }
+export { ContactItem } from './ContactItem'
 export { ContactItemActionsContainer } from '../../ContactItemActions.container'
